// src/app/models/quotationDetail.ts
export class QuotationDetail {
  EncryptedQuotationId: string;
  EncryptedQuotationItemId: string;
  EncryptedAuctionId: string;
  EncryptedQuotationDetailId: string;
  MakerName: string;
  ModelName: string;
  ModelYear: number;
  ChassisNumber: string;
  Mileage: number;
  CarImage: string;
  MaxBidPrice: string;
  BestBidPrice: string;
  ShippingPrice: string;
  ServiceCharges: string;
  InspectionPrice: string;
  CNF: string;
  EngineCC: number;
  PriceStart: number;
  PriceAverage: number;
  PriceFinish: number;
  LotNumber: string;
  AuctionHouseName: string;
  AuctionHouseId: number;
  PortId: number;
  CountryId: number;
  ShipmentTermId: number;
  ShipmentTypeId: number;
  CountryCode: string;
  ModelId: number;
  PackageId: number;
  PackageName: string;
  SentStatusId: number;
  SentStatus: string;
  SentDate: any;
  ViewDate: any;
  IsViewed: boolean;
  BidPrice: number;
  FOB: number;
  IsBooked: boolean;
  Equipment: string;
  QuotationCreatedDate: any;
  AuctionDate: any;
  ApiUniqueKey: string;
}